Simplify Badge component
