+++ lib/experimental/Information/Badge/index.stories.tsx
@@ -1,0 +1,34 @@
+import * as Icons from "@/icons/app"
+import type { Meta, StoryObj } from "@storybook/react"
+import { Badge } from "."
+
+const meta: Meta<typeof Badge> = {
+  component: Badge,
+  parameters: {
+    layout: "centered",
+  },
+  tags: ["autodocs"],
+}
+
+export default meta
+
+type Story = StoryObj<typeof Badge>
+
+export const Default: Story = {
+  args: {
+    type: "neutral",
+    icon: Icons.Placeholder,
+    size: "md",
+  },
+}
+
+export const Variants: Story = {
+  render: () => (
+    <div className="flex flex-row gap-2">
+      <Badge icon={Icons.Placeholder} size="md" />
+      <Badge icon={Icons.Check} type="positive" size="md" />
+      <Badge icon={Icons.Cross} type="critical" size="md" />
+      <Badge icon={Icons.Alert} type="warning" size="md" />
+    </div>
+  ),
+}
