--- lib/experimental/Information/Badges/exports.ts
@@ -1,2 +1,0 @@
-export * from "./EmojiBadge"
-export * from "./StatusBadge"
