+++ lib/experimental/Information/exports.tsx
@@ -1,6 +1,6 @@
 export * from "./Alert"
 export * from "./Avatars/exports"
-export * from "./Badges/exports"
+export * from "./Badge"
 export * from "./Communities/exports"
 export * from "./Counter"
 export * from "./Headers/exports"
